refactor(cleanup): drop dead legacy HTML functions and bench module
Remove production-dead code from the published surface:

- Delete parseHtmlBlock, processHtmlTokens, containsMultipleTags from
  token-cleanup.ts. All three had zero non-test importers and were
  superseded by the single-pass expandHtmlBlockNested / pairFlatHtmlTokens
  path, leaving a duplicate stack-based pairing implementation free to
  drift from the live one.
- Delete their dedicated test files and trim the cases in
  token-cleanup.test.ts that exercised them.
- Delete stream-benchmark.ts and its test; the benchmark helper had no
  production importer but still shipped in the package. Its lone consumer,
  the perf-bench dev route, now inlines the ~20-line append-parse helper.

Refresh stale doc comments that referenced the removed functions.
