+++ .agents/.plans/streaming-component-hardening/README.md
@@ -29,7 +29,7 @@ remains).
 | 005  | Prop-driven streaming batches through the rAF scheduler (#327)                               | P2       | M      | MED  | —                  | DONE   |
 | 006  | Heading-id dedup snapshots slugger occurrences (#337)                                        | P3       | M      | MED  | —                  | DONE   |
 | 007  | Streaming hot-path micro-opts: single-pass reuse array + dedupe `startsWith` (#332 residual) | P3       | S      | LOW  | —                  | DONE   |
-| 008  | Remove dead legacy HTML-pairing code + bench module                                          | P2       | S      | LOW  | —                  | TODO   |
+| 008  | Remove dead legacy HTML-pairing code + bench module                                          | P2       | S      | LOW  | —                  | DONE   |
 | 009  | Lint test files + enable type-aware promise rules                                            | P2       | M      | MED  | 008 (soft)         | TODO   |
 | 010  | Coverage thresholds + `engines` + rAF-fallback test                                          | P3       | S      | LOW  | —                  | TODO   |
 | 011  | Unify streaming token-reuse identity + generic child walk (#331, #333)                       | P2       | L      | HIGH | 003,005,007 (soft) | TODO   |
